+++ libnopegl/src/android_ctx.h
@@ -1,4 +1,5 @@
 /*
+ * Copyright 2024 Nope Forge
  * Copyright 2020-2022 GoPro Inc.
  *
  * Licensed to the Apache Software Foundation (ASF) under one
@@ -48,9 +49,6 @@ struct android_ctx {
 
     /* AHardwareBuffer */
     void (*AHardwareBuffer_describe)(const AHardwareBuffer *buffer, AHardwareBuffer_Desc *outDesc);
-
-    int has_native_imagereader_api;
-    int has_surface_texture_api;
 };
 
 int ngli_android_ctx_init(struct gpu_ctx *gpu_ctx, struct android_ctx *s);