+++ CHANGELOG.md
@@ -64,6 +64,7 @@ Versioning](https://semver.org/spec/v2.0.0.html) for `libnopegl`.
 - `%s_dimensions` uniform for 2D array and 3D images/textures, users must use
   textureSize()/imageSize() instead (`%s_dimensions` is still available for 2D
   textures)
+- `SceneInfo.files` and `SceneCfg.files`
 
 ### Removed
 - `ngl_scene_freep` function; use `ngl_scene_unrefp` instead